Created(August 2017)
Text generator is based on Markov Chains. Main data structure - Map of Lists which contain N words that are next to current(map-word) in whole text. Generation is based on a random choice of word in map, followed by Markov Chaining until required amount of words + extra words needed to meet the end of sentence.